refactor: change folder structure of project
