--- src/Chart/Chart.css
@@ -1,5 +1,0 @@
-.card {
-  width: 1200px;
-  height: 350px;
-  margin: 0 auto;
-}
