--- src/data/data.ts
@@ -1,150 +1,0 @@
-//Данные для приложения
-
-const mockData = [
-  {
-    date: '2016-02-01',
-    month: 'фев 2016',
-    indicator: 'Курс доллара',
-    value: 72,
-  },
-  {
-    date: '2016-03-02',
-    month: 'мар 2016',
-    indicator: 'Курс доллара',
-    value: 80,
-  },
-  {
-    date: '2016-04-01',
-    month: 'апр 2016',
-    indicator: 'Курс доллара',
-    value: 77,
-  },
-  {
-    date: '2016-05-02',
-    month: 'май 2016',
-    indicator: 'Курс доллара',
-    value: 78,
-  },
-  {
-    date: '2016-02-01',
-    month: 'июн 2016',
-    indicator: 'Курс доллара',
-    value: 77,
-  },
-  {
-    date: '2016-03-02',
-    month: 'июл 2016',
-    indicator: 'Курс доллара',
-    value: 76,
-  },
-  {
-    date: '2016-04-01',
-    month: 'авг 2016',
-    indicator: 'Курс доллара',
-    value: 81,
-  },
-  {
-    date: '2016-05-02',
-    month: 'сент 2016',
-    indicator: 'Курс доллара',
-    value: 82,
-  },
-
-  {
-    date: '2016-02-01',
-    month: 'фев 2016',
-    indicator: 'Курс евро',
-    value: 90,
-  },
-  {
-    date: '2016-03-02',
-    month: 'мар 2016',
-    indicator: 'Курс евро',
-    value: 88,
-  },
-  {
-    date: '2016-04-01',
-    month: 'апр 2016',
-    indicator: 'Курс евро',
-    value: 87,
-  },
-  {
-    date: '2016-05-02',
-    month: 'май 2016',
-    indicator: 'Курс евро',
-    value: 91,
-  },
-  {
-    date: '2016-02-01',
-    month: 'июн 2016',
-    indicator: 'Курс евро',
-    value: 92,
-  },
-  {
-    date: '2016-03-02',
-    month: 'июл 2016',
-    indicator: 'Курс евро',
-    value: 93,
-  },
-  {
-    date: '2016-04-01',
-    month: 'авг 2016',
-    indicator: 'Курс евро',
-    value: 89,
-  },
-  {
-    date: '2016-05-02',
-    month: 'сент 2016',
-    indicator: 'Курс евро',
-    value: 88,
-  },
-
-  {
-    date: '2016-02-01',
-    month: 'фев 2016',
-    indicator: 'Курс юаня',
-    value: 22,
-  },
-  {
-    date: '2016-03-02',
-    month: 'мар 2016',
-    indicator: 'Курс юаня',
-    value: 24,
-  },
-  {
-    date: '2016-04-01',
-    month: 'апр 2016',
-    indicator: 'Курс юаня',
-    value: 25,
-  },
-  {
-    date: '2016-05-02',
-    month: 'май 2016',
-    indicator: 'Курс юаня',
-    value: 21,
-  },
-  {
-    date: '2016-02-01',
-    month: 'июн 2016',
-    indicator: 'Курс юаня',
-    value: 23,
-  },
-  {
-    date: '2016-03-02',
-    month: 'июл 2016',
-    indicator: 'Курс юаня',
-    value: 24,
-  },
-  {
-    date: '2016-04-01',
-    month: 'авг 2016',
-    indicator: 'Курс юаня',
-    value: 26,
-  },
-  {
-    date: '2016-05-02',
-    month: 'сент 2016',
-    indicator: 'Курс юаня',
-    value: 19,
-  },
-];
