+++ src/index.css
@@ -1,0 +1,20 @@
+* {
+  padding: 0px;
+  margin: 0px;
+  border: none;
+}
+
+.container {
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.1);
+}
+
+.dashboard {
+  width: 1200px;
+  height: 350px;
+  background-color: #ffffff;
+}
