+++ src/types/types.ts
@@ -1,0 +1,14 @@
+export interface IDataItem {
+  date: string;
+  month: string;
+  indicator: string;
+  value: number;
+}
+
+export interface ITransformedData {
+  usd: IDataItem[];
+  eur: IDataItem[];
+  cny: IDataItem[];
+}
+
+export type ChartType = 'usd' | 'eur' | 'cny';
